requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==24.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.15
exceptiongroup==1.2.2
Flask==3.1.0
frozenlist==1.5.0
future==1.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.14
langchain-core==0.3.30
langchain-ollama==0.2.2
langchain-text-splitters==0.3.5
langsmith==0.2.11
line-bot-sdk==3.14.4
MarkupSafe==3.0.2
multidict==6.1.0
numpy==1.26.4
ollama==0.4.6
opencc-python-reimplemented==0.1.7
orjson==3.10.15
packaging==24.2
propcache==0.2.1
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
